A ggplot theme for Unifrog employees based on theme_grey()

A ggplot theme for Unifrog employees based on theme_grey()

Usage

unifrog_theme(
  base_size = 14,
  base_family = "Roboto",
  header_family = "Open Sans",
  base_line_size = base_size/28,
  base_rect_size = base_size/28,
  grid_col = unifrog_colors("lightgrey"),
  grid = TRUE,
  axis_col = unifrog_colors("darkgrey"),
  axis = FALSE
)

Arguments

base_size

The default font_size in pts

base_family

The default font family for text

header_family

The font family for title, axis titles, subtitle and caption.

base_line_size

Default line width

base_rect_size

Default rectangle line width

grid_col

Colour of major gridlines

grid

Should gridlines be shown? (Default = TRUE) False means hide both. X shows major X. Xx shows major and minor x. XY shows both.

axis_col

= Colour of axes

axis

show axes? TRUE/FALSE/'x'/'y'